Gallery2: Fix thumbnails of new edit photos display wrongly in Documents.
The photo is edited and save to new file in gallery,
the thumbnail of the new photo is generated and saved at the same time.
But the directory of thumbnails is not created, and the thumbnail can't save.
So the thumbnail displays incorrectly in Documents's Recent list.

Check the directory of thumbnails whether create,
if not create it at first to make the thumbnails save successfully.

Change-Id: I110a7bf568750cc9a0350d99f0a68f6b18c14fbd
CRs-Fixed: 994316